Add files via upload
This version adds dummy players, all enclosed in one object, whose money updates after you copy them and when they choose correctly. The numCorrect is tracked for each player as well. Their choices are randomly generated when the isAnswerCorrect function is called with an argument true. 
The implementation of dummy players in one object and the fact that their answers are generated in isAnswerCorrect are both inconvenient and will be changed; this is a rough draft.